+++ enclone_stuff/Cargo.toml
@@ -24,7 +24,6 @@ amino = "0.1"
 debruijn = "0.3"
 enclone_core = { path = "../enclone_core" }
 enclone_print = { path = "../enclone_print" }
-enclone_tail = { path = "../enclone_tail" }
 enclone = { path = "../enclone" }
 enclone_args = { path = "../enclone_args" }
 equiv = "0.1"
