+++ enclone_stuff/src/start.rs
@@ -21,8 +21,8 @@ use enclone::misc3::sort_tig_bc;
 use enclone_args::read_json::parse_json_annotations_files;
 use enclone_core::defs::{CloneInfo, TigData};
 use enclone_core::enclone_structs::*;
+use enclone_core::hcomp::heavy_complexity;
 use enclone_print::loupe::make_donor_refs;
-use enclone_tail::align_n::heavy_complexity;
 use equiv::EquivRel;
 use io_utils::{fwriteln, open_for_read};
 use qd::dd;
